requirements.txt
colorama==0.4.6
cryptography==42.0.5
pyodbc==5.1.0
PyYAML==6.0.2
Requests==2.32.3
pyinstaller==6.6.0
